Migrate urls config and add editable vscodeAPP
